+++ backend/API/apps/employee_routers/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Table, Column, ForeignKey, Integer, DateTime, VARCHAR, Float
+from sqlalchemy import Table, Column, ForeignKey, Integer, DateTime, VARCHAR, Float, BigInteger
 from apps.repository import Base
 
 
@@ -7,4 +7,4 @@ class employee_info(Base):
     idEmployee = Column(Integer, primary_key=True, nullable=True)
     Name = Column(VARCHAR(50), nullable=True)
     Surname = Column(VARCHAR(50), nullable=True)
-    phoneNumber = Column(Integer, nullable=True)
+    phoneNumber = Column(BigInteger, nullable=True)
